+++ OSKengine/OSKsettings.h
@@ -10,18 +10,18 @@ namespace OSK {
 
 	constexpr auto ENGINE_VERSION_STAGE_NUMERIC = 0;
 
-	constexpr auto ENGINE_VERSION_NUMERIC = 16;
+	constexpr auto ENGINE_VERSION_NUMERIC = 17;
 
-	constexpr auto ENGINE_VERSION_MINOR = "16.1";
+	constexpr auto ENGINE_VERSION_MINOR = "17*";
 
-	constexpr auto ENGINE_VERSION_ALPHANUMERIC = "Alpha 16.1";
+	constexpr auto ENGINE_VERSION_ALPHANUMERIC = "Alpha 17*";
 
 	constexpr auto ENGINE_VERSION_BUILD = "2020.10.21a";
 
-	constexpr auto ENGINE_VERSION_BUILD_NUMERIC = 20201021;
+	constexpr auto ENGINE_VERSION_BUILD_NUMERIC = 20201025;
 
-	constexpr auto ENGINE_VERSION = "0.16.1 build 2020.10.21a";
+	constexpr auto ENGINE_VERSION = "0.17* build 2020.10.25a";
 
-	constexpr auto ENGINE_VERSION_TOTAL = 16;
+	constexpr auto ENGINE_VERSION_TOTAL = 17;
 
 }