+++ OSKengine/SpriteBatch.h
@@ -10,11 +10,16 @@
 #include "AnchorEnum.h"
 #include "AnchorTextToEnum.h"
 
+#include "SpriteContainer.h"
+#include "ReservedText.h"
+
 namespace OSK {
 
 	//Clase para el renderizado 2D.
 	class OSKAPI_CALL SpriteBatch {
+		
 		friend class VulkanRenderer;
+	
 	public:
 
 		//Dibuja un sprite.
@@ -29,12 +34,37 @@ namespace OSK {
 		//	<limitAction>: establece que ocurre cuando el texto sale de la pantalla.
 		//	<sizeXlimit>: si > 0, establece, en píxeles, el límite sobre el cual se aplica <limitAction>.
 		void DrawString(const Font* fuente, const std::string& texto, const float_t& size, const Vector2& position, const Color& color = Color(1.0f), const Anchor& screenAnchor = Anchor::TOP_LEFT, const Vector4& reference = Vector4(-1.0f), const TextRenderingLimit& limitAction = TextRenderingLimit::DO_NOTHING, const float_t& sizeXlimit = 0, const float_t& limitOffset = 10);
+		
+		//Precalcula un texto reservado.
+		//Debe llamarse sólamente cuando el texto (o su posición) cambie.
+		//	<fuente>: fuente del texto.
+		//	<texto>: texto que se renderiza.
+		//	<position>: posición en la que se renderiza el texto, en píxeles desde el ancla (<screenAnchor>), por defecto desde la esquina inferior izquierda de la pantalla.
+		//	<color>: color del texto.
+		//	<screenAnchor>: establece el origen de coordenadas desde el que se calcula la posición del texto.
+		//	<limitAction>: establece que ocurre cuando el texto sale de la pantalla.
+		//	<sizeXlimit>: si > 0, establece, en píxeles, el límite sobre el cual se aplica <limitAction>.
+		void PrecalculateText(const Font* fuente, const ReservedText& texto, const float_t& size, const Vector2& position, const Color& color = Color(1.0f), const Anchor& screenAnchor = Anchor::TOP_LEFT, const Vector4& reference = Vector4(-1.0f), const TextRenderingLimit& limitAction = TextRenderingLimit::DO_NOTHING, const float_t& sizeXlimit = 0, const float_t& limitOffset = 10);
+
+		//Dibuja un texto reservado.
+		//	<texto>: texto.
+		void DrawString(const ReservedText& texto);
+
+		//Precalcula y renderiza un texto reservado.
+		//	<fuente>: fuente del texto.
+		//	<texto>: texto que se renderiza.
+		//	<position>: posición en la que se renderiza el texto, en píxeles desde el ancla (<screenAnchor>), por defecto desde la esquina inferior izquierda de la pantalla.
+		//	<color>: color del texto.
+		//	<screenAnchor>: establece el origen de coordenadas desde el que se calcula la posición del texto.
+		//	<limitAction>: establece que ocurre cuando el texto sale de la pantalla.
+		//	<sizeXlimit>: si > 0, establece, en píxeles, el límite sobre el cual se aplica <limitAction>.
+		void DrawString(const Font* fuente, const ReservedText& texto, const float_t& size, const Vector2& position, const Color& color = Color(1.0f), const Anchor& screenAnchor = Anchor::TOP_LEFT, const Vector4& reference = Vector4(-1.0f), const TextRenderingLimit& limitAction = TextRenderingLimit::DO_NOTHING, const float_t& sizeXlimit = 0, const float_t& limitOffset = 10);
 
 		//Vacía el spriteBatch.
 		void Clear();
 
 	private:
-		
+
 		SpriteBatch();
 
 		Vector2 GetTextPosition(const Vector2& position, const Vector2& textSize, const Anchor& anchor, const AnchorTextTo& to, const Vector4& reference = Vector4(0.0f)) const;
@@ -46,7 +76,7 @@ namespace OSK {
 		//vs
 		//2.5
 		//190
-		std::deque<Sprite> spritesToDraw{};
+		std::deque<SpriteContainer> spritesToDraw{};
 
 	};