+++ OSKengine/resource.h
@@ -1,12 +1,15 @@
 //{{NO_DEPENDENCIES}}
-// Microsoft Visual C++ generated include file.
-// Used by OSKengine.rc
+// Archivo de inclusión generado de Microsoft Visual C++.
+// Usado por OSKengine.rc
+//
+#define IDI_ICON1                       101
+#define IDI_ICON2                       102
 
-// Valores predeterminados siguientes para nuevos objetos
+// Next default values for new objects
 // 
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NEXT_RESOURCE_VALUE        101
+#define _APS_NEXT_RESOURCE_VALUE        103
 #define _APS_NEXT_COMMAND_VALUE         40001
 #define _APS_NEXT_CONTROL_VALUE         1001
 #define _APS_NEXT_SYMED_VALUE           101